Beginning of French translations for ATM9
Hello everyone,

This pull request includes the initial French translations of ATM9, with the available elements. I translated and then reworked the sentences to make them much more understandable in French. The meaning of the sentences is identical to the original version.I also want to emphasize that I will translate the mods of ATM9 that are integral to quests, tips, etc., to ensure that the ATM9 translations correspond to the mods highlighted by ATM9 and that are sure not to be deleted, so that the gaming experience is as optimal as possible.

I would like to thank you for your time. Looking forward to continuing this, thank you for your trust!